% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ssh_functions.R
\name{args_2_string}
\alias{args_2_string}
\title{Convert list to string}
\usage{
args_2_string(args)
}
\value{
A fun_argument string
}
\description{
Convert list to string
}
\author{
Giovanni Laudanno
}
