% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.R
\name{branching_times}
\alias{branching_times}
\title{Extrapolate tidy branching times from phylogeny}
\usage{
branching_times(phylogeny)
}
\description{
Extrapolate tidy branching times from phylogeny
}
\author{
Giovanni Laudanno
}
