% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{build_description_file}
\alias{build_description_file}
\title{Adds all require dependencies to the DESCRIPTION file}
\usage{
build_description_file(project_name, ...)
}
\arguments{
\item{project_name}{the name of the project}
}
\description{
Adds all require dependencies to the DESCRIPTION file
}
\author{
Giovanni Laudanno
}
