% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{check_jobs}
\alias{check_jobs}
\title{Check jobs on cluster}
\usage{
check_jobs(account = "p274829", session = NA)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
list with job ids, job info and sshare
}
\description{
Check jobs on cluster
}
\author{
Giovanni Laudanno, Pedro Neves
}
