% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{close_connection}
\alias{close_connection}
\title{Close connection}
\usage{
close_connection(connection)
}
\arguments{
\item{connection}{a ssh connection}
}
\value{
nothing
}
\description{
Close a connection for a given account
}
\author{
Giovanni Laudanno
}
