% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{close_jobs}
\alias{close_jobs}
\title{Close jobs on cluster}
\usage{
close_jobs(account = "p274829")
}
\arguments{
\item{account}{a peregrine account}
}
\value{
list with job ids, job info and sshare (all empty)
}
\description{
Close jobs on cluster
}
\author{
Giovanni Laudanno
}
