% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{close_session}
\alias{close_session}
\title{Close session}
\usage{
close_session(session)
}
\arguments{
\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Close a session for a given account
}
\author{
Giovanni Laudanno
}
