% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_functions.R
\name{delete_folder_structure}
\alias{delete_folder_structure}
\title{Delete folder structure}
\usage{
delete_folder_structure(
  account = "p274829",
  projects_folder_name = "Projects",
  disk = "D",
  session = NA,
  drive = FALSE
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Delete folder structure
}
\author{
Giovanni Laudanno
}
