% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ssh_functions.R
\name{download_results}
\alias{download_results}
\title{Download the results to the results folder of the project}
\usage{
download_results(
  local_projects_folder = "D:/Projects",
  remote_projects_folder = file.path("/home", account),
  project_name = "sls",
  delete_on_cluster = FALSE,
  account = "p274829",
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Download the results to the results folder of the project
}
\author{
Giovanni Laudanno
}
