% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive_functions.R
\name{drive_dir_create}
\alias{drive_dir_create}
\title{Create a directory on drive}
\usage{
drive_dir_create(dir)
}
\arguments{
\item{dir}{a directory}
}
\value{
nothing
}
\description{
Create a directory on drive
}
\author{
Giovanni Laudanno
}
