% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive_functions.R
\name{drive_list_files}
\alias{drive_list_files}
\title{List drive files}
\usage{
drive_list_files(dir = jap::default_projects_folder())
}
\arguments{
\item{dir}{a directory}
}
\value{
List of files
}
\description{
List drive files
}
\author{
Giovanni Laudanno
}
