% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{file_path}
\alias{file_path}
\title{Like file.path, but cooler}
\usage{
file_path(..., fsep = .Platform$file.sep)
}
\arguments{
\item{...}{additional arguments}

\item{fsep}{path separator for the OS}
}
\description{
Like file.path, but cooler
}
