% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_functions.R
\name{fix_documentation}
\alias{fix_documentation}
\title{Fix End of Line documentation problem}
\usage{
fix_documentation()
}
\value{
nothing
}
\description{
Fix End of Line documentation problem. Don't forget to refresh
your Git tab in Rstudio afterwards.
}
\author{
Giovanni Laudanno
}
