% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_functions.R
\name{folder_structure}
\alias{folder_structure}
\title{Specify subfolder structure}
\usage{
folder_structure()
}
\value{
nothing
}
\description{
Specify subfolder structure
}
\author{
Giovanni Laudanno
}
