% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{get_cluster_address}
\alias{get_cluster_address}
\title{Get peregrine cluster address}
\usage{
get_cluster_address(account = jap::your_account())
}
\arguments{
\item{account}{a peregrine account}
}
\value{
peregrine cluster address
}
\description{
Get peregrine cluster address
}
\author{
Giovanni Laudanno
}
