% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{get_function_list}
\alias{get_function_list}
\title{Get function list}
\usage{
get_function_list(project_name, my_github = "Giappo")
}
\value{
function list
}
\description{
NOT WORKING YET
}
\author{
Giovanni Laudanno
}
