% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_functions.R
\name{get_local_project_folder}
\alias{get_local_project_folder}
\title{Find local project folder}
\usage{
get_local_project_folder(
  project_name,
  projects_folder_name = jap::default_projects_folder(),
  home_dir = jap::default_home_dir()
)
}
\arguments{
\item{project_name}{the name of the project}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{home_dir}{path to the home directory. On windows usually is
a disk like C or D. On mac is always ~.}
}
\description{
Find local project folder
}
