% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_functions.R
\name{git_clone}
\alias{git_clone}
\title{Clone a repo}
\usage{
git_clone(github_name, github_repo, ...)
}
\arguments{
\item{github_name}{the github account name where the repo is}

\item{github_repo}{a github repository}

\item{...}{additional arguments to be passed to \link{find_github_folder}}
}
\description{
Clone a repo
}
\author{
Giovanni Laudanno
}
