% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_functions.R
\name{git_pull}
\alias{git_pull}
\title{Pull a repo}
\usage{
git_pull(github_name, github_repo)
}
\arguments{
\item{github_name}{the github account name where the repo is}

\item{github_repo}{a github repository}
}
\description{
Pull a repo
}
\author{
Giovanni Laudanno
}
