% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{install_package}
\alias{install_package}
\title{Install and load a package}
\usage{
install_package(package_name, github_name = NA)
}
\value{
nothing
}
\description{
Install and load a package
}
\author{
Giovanni Laudanno
}
