% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlibrary_functions.R
\name{install_richel}
\alias{install_richel}
\title{Receive the word of our lord JC directly on your PC}
\usage{
install_richel()
}
\value{
nothing
}
\description{
Receive the word of our lord JC directly on your PC
}
\author{
Giovanni Laudanno, Richel Bilderbeek
}
