% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{is_session_open}
\alias{is_session_open}
\title{Check if session is open or not}
\usage{
is_session_open(session)
}
\arguments{
\item{session}{a ssh session}
}
\value{
TRUE or FALSE
}
\description{
Check if session is open
}
\author{
Giovanni Laudanno
}
