% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_functions.R
\name{list_githubs}
\alias{list_githubs}
\title{Open github folder}
\usage{
list_githubs(...)
}
\arguments{
\item{...}{additional arguments to be passed to \link{find_github_folder}}
}
\description{
Open github folder
}
\author{
Giovanni Laudanno
}
