% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_params_doc.R
\name{nee_params_doc}
\alias{nee_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
nee_params_doc(lambda, mu, n, t, tbar)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{n}{number of species}

\item{t}{time}

\item{tbar}{time}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
}
