% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{one_minus_pt}
\alias{one_minus_pt}
\title{1 - Pt}
\usage{
one_minus_pt(lambda, mu, t)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{t}{time}
}
\value{
1 - pt
}
\description{
Nee's function: 1 - pt
}
\author{
Giovanni Laudanno
}
