% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{one_minus_ut}
\alias{one_minus_ut}
\title{1 - ut}
\usage{
one_minus_ut(lambda, mu, t)
}
\arguments{
\item{lambda}{the sympatric speciation rate.}

\item{mu}{the extinction rate.}

\item{t}{time}
}
\value{
1 - ut
}
\description{
Nee's function: 1 - ut
}
\author{
Giovanni Laudanno
}
