% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{open_connection}
\alias{open_connection}
\title{Open connection}
\usage{
open_connection(account = "p274829")
}
\arguments{
\item{account}{a peregrine account}
}
\value{
the connection
}
\description{
Open a connection for a given account
}
\author{
Giovanni Laudanno
}
