% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_functions.R
\name{open_github_folder}
\alias{open_github_folder}
\title{Open github folder}
\usage{
open_github_folder(...)
}
\arguments{
\item{...}{additional arguments to be passed to \link{find_github_folder}}
}
\description{
It opens your github folder.
}
\author{
Giovanni Laudanno
}
