% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_functions.R
\name{open_github_project}
\alias{open_github_project}
\title{Open github project}
\usage{
open_github_project(github_repo, ...)
}
\arguments{
\item{github_repo}{a github repository}

\item{...}{additional arguments to be passed to \link{find_github_folder}}
}
\description{
It opens a github project from your github folder.
To specify your github folder see \link{find_github_folder}.
}
\author{
Giovanni Laudanno
}
