% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{open_session}
\alias{open_session}
\title{Open session}
\usage{
open_session(account = jap::your_account())
}
\arguments{
\item{account}{a peregrine account}
}
\value{
the session
}
\description{
Open a session for a given account
}
\author{
Giovanni Laudanno
}
