% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{p_t}
\alias{p_t}
\title{Pt}
\usage{
p_t(lambda, mu, t)
}
\arguments{
\item{lambda}{the sympatric speciation rate.}

\item{mu}{the extinction rate.}

\item{t}{time}
}
\value{
pt
}
\description{
Nee's function: pt
}
\author{
Giovanni Laudanno
}
