% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{pars_transform_back}
\alias{pars_transform_back}
\title{Transform parameters back}
\usage{
pars_transform_back(pars_transformed)
}
\arguments{
\item{pars_transformed}{parameters already transformed}
}
\value{
the original parameters
}
\description{
Transform parameters back according to x = y / (1 + y)
}
\details{
This is not to be called by the user.
}
