% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{pars_transform_forward}
\alias{pars_transform_forward}
\title{Transform parameters}
\usage{
pars_transform_forward(pars)
}
\arguments{
\item{pars}{parameters}
}
\value{
transformed parameters
}
\description{
Transform parameters according to y = x / (1 + x)
}
\details{
This is not to be called by the user.
}
