% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{path_2_file.path}
\alias{path_2_file.path}
\title{Convert a path to the equivalent file.path call}
\usage{
path_2_file.path(path)
}
\arguments{
\item{path}{a path}
}
\description{
Convert a path to the equivalent file.path call
}
