% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.R
\name{plot_ltt}
\alias{plot_ltt}
\title{Ltt plot for a tree}
\usage{
plot_ltt(tree, present_is_zero = FALSE)
}
\arguments{
\item{present_is_zero}{Set to TRUE if you want the present time to be zero}
}
\description{
Ltt plot for a tree
}
\author{
Giovanni Laudanno
}
