% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Just plot a matrix without rotating it}
\usage{
plot_matrix(mat, logs = TRUE, low_triangular = FALSE)
}
\arguments{
\item{logs}{do you want to plot in log scale?}

\item{low_triangular}{do you want to plot only the low triangular?}
}
\description{
Just plot a matrix without rotating it
}
