% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{pn}
\alias{pn}
\title{Pn}
\usage{
pn(lambda, mu, t, n)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{t}{time}

\item{n}{number of species}
}
\value{
pn
}
\description{
Nee's function: pn
}
\author{
Giovanni Laudanno
}
