% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{pn_bar}
\alias{pn_bar}
\title{Pn accounting for extinctions after the shifts}
\usage{
pn_bar(lambda, mu, t, n, tbar = 0)
}
\arguments{
\item{lambda}{the sympatric speciation rate.}

\item{mu}{the extinction rate.}

\item{t}{time}
}
\value{
pn times probability of extinction for n-1 species after the shift
}
\description{
Combine pn from Nee et al. and imposes the extinction
 before the present of all species not visible in the phylogeny
}
\author{
Giovanni Laudanno
}
