% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{print_from_global}
\alias{print_from_global}
\title{Print a variable, but not from test environment}
\usage{
print_from_global(var = "seed")
}
\arguments{
\item{var}{a variable, to be expressed as string}
}
\description{
Print a variable, but not from test environment
}
