% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{pt}
\alias{pt}
\title{Pt}
\usage{
pt(lambda, mu, t)
}
\arguments{
\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{t}{time}
}
\value{
pt
}
\description{
Nee's function: pt
}
\author{
Giovanni Laudanno
}
