% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.R
\name{read_phylo}
\alias{read_phylo}
\title{Open tree file}
\usage{
read_phylo(tree_file)
}
\arguments{
\item{tree_file}{a tree file in format nexus or .tree}
}
\description{
Open tree file
}
\author{
Giovanni Laudanno
}
