% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_functions.R
\name{read_phylo_node_bars}
\alias{read_phylo_node_bars}
\title{Open tree file to retrieve node bars info}
\usage{
read_phylo_node_bars(tree_file)
}
\arguments{
\item{tree_file}{a tree file in format nexus or .tree}
}
\description{
Open tree file to retrieve node bars info
}
\author{
Giovanni Laudanno
}
