% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ssh_functions.R
\name{remote_dir.create}
\alias{remote_dir.create}
\title{Create a directory on cluster}
\usage{
remote_dir.create(dir, account = "p274829", session = NA)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Create a directory on cluster
}
\author{
Giovanni Laudanno
}
