% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_dir.exists}
\alias{remote_dir.exists}
\title{Check if folder exist on cluster}
\usage{
remote_dir.exists(dir, account = jap::your_account(), session = NA)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Check if folder exist on cluster
}
\author{
Giovanni Laudanno
}
