% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_dir.remove}
\alias{remote_dir.remove}
\title{Remove a directory}
\usage{
remote_dir.remove(dir, account = "p274829", session = NA)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Remove a directory
}
\author{
Giovanni Laudanno
}
