% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_dir_create}
\alias{remote_dir_create}
\title{Create a directory on cluster}
\usage{
remote_dir_create(dir, account = jap::your_account(), session = NA)
}
\arguments{
\item{dir}{a directory}

\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Create a directory on cluster
}
\author{
Giovanni Laudanno
}
