% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_dir_remove}
\alias{remote_dir_remove}
\title{Remove a directory}
\usage{
remote_dir_remove(
  path_to_dir,
  account = jap::your_account(),
  cluster_folder = jap::default_cluster_folder(),
  session = NA
)
}
\arguments{
\item{path_to_dir}{path to the directory location, from `default_cluster_folder()`}

\item{account}{a peregrine account}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Remove a directory
}
\author{
Giovanni Laudanno
}
