% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_file_exists}
\alias{remote_file_exists}
\title{Check if file exist on cluster}
\usage{
remote_file_exists(
  path_to_file,
  account = jap::your_account(),
  cluster_folder = jap::default_cluster_folder(),
  session = NA
)
}
\arguments{
\item{path_to_file}{path to the target file from \code{default_cluster_folder()}}

\item{account}{a peregrine account}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Check if file exist on cluster
}
\author{
Giovanni Laudanno
}
