% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ssh_functions.R
\name{remote_install_github}
\alias{remote_install_github}
\title{Install packages for a given project}
\usage{
remote_install_github(
  github_name = NA,
  package_name,
  must_sleep = TRUE,
  account = "p274829",
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Install packages for a given project
}
\author{
Giovanni Laudanno
}
