% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_list.files}
\alias{remote_list.files}
\title{List remote files}
\usage{
remote_list.files(dir, full.names = TRUE, session = NA)
}
\arguments{
\item{dir}{a directory}

\item{session}{a ssh session}
}
\value{
List of files
}
\description{
List remote files
}
\author{
Giovanni Laudanno
}
