% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_list_files}
\alias{remote_list_files}
\title{List remote files}
\usage{
remote_list_files(
  dir = jap::default_projects_folder(),
  cluster_folder = jap::default_cluster_folder(),
  account = jap::your_account(),
  session = NA
)
}
\arguments{
\item{dir}{a directory}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
List of files
}
\description{
List remote files
}
\author{
Giovanni Laudanno
}
