% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlibrary_functions.R
\name{remove_package}
\alias{remove_package}
\title{Remove a package (and lock file)}
\usage{
remove_package(package_name)
}
\arguments{
\item{package_name}{the name of the package}
}
\value{
nothing
}
\description{
Remove a package (and lock file)
}
\author{
Giovanni Laudanno
}
