% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{run_function_from_file}
\alias{run_function_from_file}
\title{run a function from file}
\usage{
run_function_from_file(args_file)
}
\arguments{
\item{args_file}{the file containing the parameters}
}
\value{
nothing
}
\description{
run a function from file
}
\author{
Giovanni Laudanno
}
