% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{run_pirouette_example}
\alias{run_pirouette_example}
\title{run pirouette example}
\usage{
run_pirouette_example(
  example_no,
  project_name = get_pkg_name(),
  account = "p274829",
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
run pirouette example
}
\author{
Giovanni Laudanno
}
