% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{unlock_package}
\alias{unlock_package}
\title{Unlock package when you cannot git add anymore}
\usage{
unlock_package(package_name)
}
\value{
nothing
}
\description{
Unlock package when you cannot git add anymore
}
\author{
Giovanni Laudanno
}
