% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{ut}
\alias{ut}
\title{ut}
\usage{
ut(lambda, mu, t)
}
\arguments{
\item{lambda}{the sympatric speciation rate.}

\item{mu}{the extinction rate.}

\item{t}{time}
}
\value{
ut
}
\description{
Nee's function: ut
}
\author{
Giovanni Laudanno
}
