% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ssh_functions.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Create folder structure}
\usage{
create_folder_structure(
  account = "p274829",
  local_projects_folder = "D:/Projects",
  remote_projects_folder = file.path("/home", account),
  project_name = "sls",
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Create folder structure
}
\author{
Giovanni Laudanno
}
