% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_functions.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Create folder structure}
\usage{
create_folder_structure(
  account = jap::your_account(),
  projects_folder_name = "Projects",
  home_dir = jap::default_home_dir(),
  cluster_folder = "home",
  project_name = NA,
  session = NA,
  drive = FALSE
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Create folder structure
}
\author{
Giovanni Laudanno
}
