% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  account,
  age,
  session,
  lambda,
  mu,
  projects_folder_name,
  project_name,
  t,
  message,
  verbose
)
}
\arguments{
\item{account}{a peregrine account}

\item{age}{the age of the phylogeny}

\item{session}{a ssh session}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{project_name}{the name of the project}

\item{t}{time}

\item{message}{a message to be print}

\item{verbose}{choose if you want to print the output or not}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
}
