% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(a_abstol, a_reltol, atol, abstol, account_name, after,
  age, alpha, alpha0, b, brts, brts_precision, changeloglikifnoconv,
  checked_functions, colormap, cond, data_folder, debug_mode, empty_mat,
  empty_pp, empty_qq, eq, expr, fortran, func, function_name,
  function_names, functions_names, idparsfix, idparsopt, initparsopt,
  init_n_lineages, initprobs, init_probs, input_trees_path, input_xml,
  interval_max, interval_min, iterations, k, lambda, lambda_limit,
  likelihood, l_matrix, loglik_function, loglik_functions, logs,
  log_nu_mat, log_q_mat, lx, lx0, lx_top, m1, m2, m1_mat, m2_mat, mm,
  matrices, matrix, matrix_a, matrix_builder, max_iter, max_k,
  max_number_of_species, max_repetitions, max_sims, maxit, maxiter,
  mbd_lambda, message, methode, minimum_multiple_births, missnumspec,
  models, mu, mu_limit, mutation_rate, n_0, n_0s, n_species, n_sims,
  n_steps, n_subs, nu, nu_limit, nu_matrix, optimmethod, optim_ids, pp,
  pp2, parms, parmsvec, params, pars, pars_transform, pars_transformed,
  parsfix, parsvec, pc, pool, precision, print_errors, printit,
  project_folder, pvec, q, qq, qq2, q_t, q_vector, q_threshold,
  quantiles_choice, qvec, recursive, rtol, reltol, res, results,
  results_folder, rhs_function, runmod, s, safety_checks, saveit, seed,
  sample_interval, sequence_length, sim, sim_pars, sim_phylo, start_pars,
  starting_seed, subsamp, sum_probs_1, sum_probs_2, t, tcrit, times, tvec,
  t1, t2, t_0, t_0s, tips_interval, time_interval, tol, tr,
  transition_matrix, trparsfix, trparsopt, true_pars, values, vector,
  verbose, x, x_name, x_splits, y, y_name, y_splits, z, z_name)
}
\arguments{
\item{a_abstol}{something}

\item{a_reltol}{something}

\item{atol}{absolute error tolerance for
the numerical integration using deSolve.}

\item{abstol}{absolute error tolerance for
the numerical integration using deSolve.}

\item{account_name}{something}

\item{after}{something}

\item{age}{the age of the tree.}

\item{alpha}{something}

\item{alpha0}{something}

\item{b}{the number of simultaneous births on a given branching time.}

\item{brts}{A set of branching times of a phylogeny.}

\item{brts_precision}{set the level of approximation on the branching times.}

\item{changeloglikifnoconv}{something}

\item{checked_functions}{something}

\item{colormap}{something}

\item{cond}{sets the conditioning
\itemize{
 \item cond = 0: no conditioning;
 \item cond = 1: conditioning on stem or crown age
  and non-extinction of the phylogeny;
 \item cond = 2: like cond = 1 but forces at least one visible multiple
 event in the phylogeny
}}

\item{data_folder}{The data folder insider the project folder.}

\item{debug_mode}{If TRUE allows to run even when there are errors and
expose them.}

\item{empty_mat}{an empty matrix}

\item{empty_pp}{empty pp matrix. See \code{pp}.}

\item{empty_qq}{empty qq matrix. See \code{qq}.}

\item{eq}{the equation approach you want to use. It can be "q_eq" for the
Q-equation of "p_eq" for the P-equation.}

\item{expr}{and expression}

\item{fortran}{Set it to TRUE if you want to use FORTRAN routine.}

\item{func}{a function}

\item{function_name}{function name}

\item{function_names}{function names}

\item{functions_names}{function names}

\item{idparsfix}{something}

\item{idparsopt}{the ids of the parameters that must be optimized.
The ids are defined as follows:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability_
}}

\item{initparsopt}{something}

\item{init_n_lineages}{the number of lineages at time equals zero.}

\item{initprobs}{initial probabilities}

\item{init_probs}{initial probabilities}

\item{input_trees_path}{Path to \code{.trees} file.}

\item{input_xml}{something}

\item{interval_max}{something}

\item{interval_min}{something}

\item{iterations}{something}

\item{k}{the number of visible species in the phylogeny at a given time.}

\item{lambda}{the sympatric speciation rate.}

\item{lambda_limit}{Upper limit to lambda estimations.}

\item{likelihood}{the likelihood}

\item{l_matrix}{the l-table}

\item{loglik_function}{the loglik function}

\item{loglik_functions}{the loglik_functions you want to use}

\item{logs}{something}

\item{log_nu_mat}{logarithmic nu-matrix, used for condprob}

\item{log_q_mat}{logarithmic q-matrix, used for condprob}

\item{lx}{it is the number of ODEs considered for the computation.}

\item{lx0}{something}

\item{lx_top}{maximum lx to produce pre-calculated matrices for condprob}

\item{m1}{a matrix keeping track of columns}

\item{m2}{a matrix keeping track of rows}

\item{m1_mat}{a matrix keeping track of columns}

\item{m2_mat}{a matrix keeping track of rows}

\item{mm}{indexes of missing species}

\item{matrices}{matrices}

\item{matrix}{a matrix}

\item{matrix_a}{The A matrix from the theory that represents the ODE system.}

\item{matrix_builder}{function used to build the transition matrix.
Default option is \code{hyper_a_hanno}}

\item{max_iter}{Sets the maximum number of iterations in the optimization}

\item{max_k}{something}

\item{max_number_of_species}{something}

\item{max_repetitions}{something}

\item{max_sims}{something}

\item{maxit}{maximum number of subplex iterations}

\item{maxiter}{maximum number of subplex iterations}

\item{mbd_lambda}{something}

\item{message}{a message}

\item{methode}{specifies how the integration must be performed:
\itemize{
  \item \code{lsodes}: use \code{lsodes} and \code{deSolve::ode}
  \item \code{ode45}: use \code{ode45} and \code{deSolve::ode}
  \item \code{lsoda}: use \code{lsoda} and \code{deSolve::ode}
}}

\item{minimum_multiple_births}{minimum amount of multiple births
that have to be present in the simulated phylogeny.}

\item{missnumspec}{The number of species that are in the clade,
but missing in the phylogeny.}

\item{models}{the models you want to use to define the likelihood}

\item{mu}{the extinction rate.}

\item{mu_limit}{Upper limit to mu estimations.}

\item{mutation_rate}{something}

\item{n_0}{the number of lineages at time equals zero.}

\item{n_0s}{starting number of lineages for all the clades}

\item{n_species}{number of species.}

\item{n_sims}{number of simulations}

\item{n_steps}{something}

\item{n_subs}{something}

\item{nu}{the multiple allopatric speciation trigger rate.}

\item{nu_limit}{Upper limit to nu estimations.}

\item{nu_matrix}{matrix for the nu component. Component {m,n} is equal to
choose(n, m - n) x q ^ (m - n) x (1 - q) ^ (2 * n - m)}

\item{optimmethod}{optimization routine: choose between subplex and simplex}

\item{optim_ids}{ids of the parameters you want to optimize.}

\item{pp}{the matrix p_{n1, n2}}

\item{pp2}{the matrix p_{n1, n2}, with an empty frame}

\item{parms}{some parameters to pass on the ode function}

\item{parmsvec}{parameters to pass to the FORTRAN scripts}

\item{params}{transition matrix for the rhs of the ode system.}

\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}

\item{pars_transform}{something}

\item{pars_transformed}{parameters of the likelihood functions, transformed
according to y = x / (1 + x)}

\item{parsfix}{the values of the parameters that should not be optimized.}

\item{parsvec}{vector of paramters for the FORTRAN subroutine.}

\item{pc}{conditional probability}

\item{pool}{the pool of living species at a given time}

\item{precision}{something}

\item{print_errors}{something}

\item{printit}{something}

\item{project_folder}{the folder when you want to save data and results}

\item{pvec}{a vector of probabilities P}

\item{q}{the single-lineage speciation probability at a triggered event.}

\item{qq}{the matrix q_{m1, m2}}

\item{qq2}{the matrix q_{m1, m2}, with an empty frame}

\item{q_t}{the q_vector in time}

\item{q_vector}{the \code{Q} vector from
'Etienne et al. 2012 - Proc. R. Soc. B'.}

\item{q_threshold}{adds a threshold for the evaluation of q. This is due
because you never want \code{q} to actually be equal to zero or one.}

\item{quantiles_choice}{something}

\item{qvec}{a vector of probabilities Q}

\item{recursive}{something}

\item{rtol}{relative error tolerance for
the numerical integration using deSolve.}

\item{reltol}{relative error tolerance for
the numerical integration using deSolve.}

\item{res}{something}

\item{results}{mle results}

\item{results_folder}{The results folder insider the project folder.}

\item{rhs_function}{a function for the right hand side of the differential
equation}

\item{runmod}{FORTRAN subroutine's option}

\item{s}{the seed}

\item{safety_checks}{TRUE if you want the parameters to be limited to a
realistic, but not infinite, parameter space}

\item{saveit}{do you want to save the results?}

\item{seed}{the seed}

\item{sample_interval}{something}

\item{sequence_length}{something}

\item{sim}{the results of a sim run}

\item{sim_pars}{vector of parameters:
\itemize{
  \item id == 1 corresponds to lambda (speciation rate)
  \item id == 2 corresponds to mu (extinction rate)
  \item id == 3 corresponds to nu (multiple speciation trigger rate)
  \item id == 4 corresponds to q (single-lineage speciation probability)
}}

\item{sim_phylo}{something}

\item{start_pars}{starting parameter values for the ml process.}

\item{starting_seed}{the seed to start from}

\item{subsamp}{something}

\item{sum_probs_1}{vector of probability sums}

\item{sum_probs_2}{vector of probability sums}

\item{t}{time}

\item{tcrit}{the time you don't want to exceed in the integration}

\item{times}{vector of times}

\item{tvec}{vector of times}

\item{t1}{something}

\item{t2}{something}

\item{t_0}{starting time}

\item{t_0s}{starting time for each clade}

\item{tips_interval}{sets tips boundaries constrain on simulated dataset.
It works only if cond == 1, otherwise it must be set to c(0, Inf).}

\item{time_interval}{a time interval}

\item{tol}{something}

\item{tr}{a phylogeny of class \code{phylo}}

\item{transition_matrix}{something}

\item{trparsfix}{something}

\item{trparsopt}{something}

\item{true_pars}{true parameter values when running the ml process.}

\item{values}{something}

\item{vector}{a vector}

\item{verbose}{choose if you want to print the output or not}

\item{x}{something}

\item{x_name}{something}

\item{x_splits}{something}

\item{y}{something}

\item{y_name}{something}

\item{y_splits}{something}

\item{z}{something}

\item{z_name}{something}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
  use of this function by Richel J.C. Bilderbeek
}
