% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  account,
  age,
  cluster_folder,
  delete_on_cluster,
  dir,
  drive,
  file,
  function_name,
  fun_arguments,
  github_name,
  github_repo,
  github_folder_name,
  home_dir,
  lambda,
  message,
  mu,
  package_name,
  path,
  projects_folder_name,
  project_name,
  session,
  subfolder,
  t,
  verbose
)
}
\arguments{
\item{account}{a peregrine account}

\item{age}{the age of the phylogeny}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{delete_on_cluster}{do you want to delete the files from the cluster
after the download?}

\item{dir}{a directory}

\item{drive}{do you want to use google drive to store the data of your
prokects? It can be either TRUE or FALSE}

\item{file}{a file}

\item{function_name}{the name of the function}

\item{fun_arguments}{the arguments of the function passed either as
list or as string (in the format 'arg1 = x, arg2 = y')}

\item{github_name}{the github account name where the repo is}

\item{github_repo}{a github repository}

\item{github_folder_name}{name of the local folder containing all the
github repos}

\item{home_dir}{path to the home directory. On windows usually is
a disk like C or D. On mac is always ~.}

\item{lambda}{speciation rate}

\item{message}{a message to be print}

\item{mu}{extinction rate}

\item{package_name}{the name of the package}

\item{path}{a path}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{project_name}{the name of the project}

\item{session}{a ssh session}

\item{subfolder}{a subfolder of the project folder}

\item{t}{time}

\item{verbose}{choose if you want to print the output or not}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
}
