% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_functions.R
\name{delete_folder_structure}
\alias{delete_folder_structure}
\title{Delete folder structure}
\usage{
delete_folder_structure(
  account = jap::your_account(),
  projects_folder_name = jap::default_projects_folder(),
  home_dir = jap::default_home_dir(),
  cluster_folder = jap::default_cluster_folder(),
  drive = jap::default_drive_choice(),
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{home_dir}{path to the home directory. On windows usually is
a disk like C or D. On mac is always ~.}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{drive}{do you want to use google drive to store the data of your
prokects? It can be either TRUE or FALSE}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Delete folder structure
}
\author{
Giovanni Laudanno
}
