% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_structure_functions.R
\name{delete_folder_structure}
\alias{delete_folder_structure}
\title{Delete folder structure}
\usage{
delete_folder_structure(
  account = jap::your_account(),
  projects_folder_name = jap::default_projects_folder(),
  home_dir = jap::default_home_dir(),
  cluster_folder = "home",
  session = NA,
  drive = FALSE
)
}
\arguments{
\item{account}{a peregrine account}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Delete folder structure
}
\author{
Giovanni Laudanno
}
