% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{download_project_folder}
\alias{download_project_folder}
\title{Download the entire project folder}
\usage{
download_project_folder(
  function_name = NA,
  project_name = NA,
  projects_folder_name = jap::default_projects_folder(),
  account = jap::your_account(),
  cluster_folder = jap::default_cluster_folder(),
  home_dir = jap::default_home_dir(),
  delete_on_cluster = FALSE,
  drive = jap::default_drive_choice(),
  session = NA
)
}
\arguments{
\item{function_name}{the name of the function}

\item{project_name}{the name of the project}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{account}{a peregrine account}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{home_dir}{path to the home directory. On windows usually is
a disk like C or D. On mac is always ~.}

\item{delete_on_cluster}{do you want to delete the files from the cluster
after the download?}

\item{drive}{do you want to use google drive to store the data of your
prokects? It can be either TRUE or FALSE}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Download the entire project folder
}
\author{
Giovanni Laudanno
}
