% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{download_subfolder}
\alias{download_subfolder}
\title{Download the results to the results folder of the project}
\usage{
download_subfolder(
  subfolder = "results",
  projects_folder_name = "Projects",
  disk = "D",
  project_name = "sls",
  delete_on_cluster = FALSE,
  account = jap::your_account(),
  cluster_folder = "home",
  session = NA,
  drive = FALSE
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Download the results to the results folder of the project
}
\author{
Giovanni Laudanno
}
