% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{download_subfolder}
\alias{download_subfolder}
\title{Download the results to the results folder of the project}
\usage{
download_subfolder(
  subfolder = "results",
  projects_folder_name = jap::default_projects_folder(),
  home_dir = jap::default_home_dir(),
  cluster_folder = jap::default_cluster_folder(),
  project_name = "sls",
  delete_on_cluster = FALSE,
  account = jap::your_account(),
  drive = jap::default_drive_choice(),
  session = NA
)
}
\arguments{
\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{project_name}{the name of the project}

\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Download the results to the results folder of the project
}
\author{
Giovanni Laudanno
}
