% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_functions.R
\name{find_github_folder}
\alias{find_github_folder}
\title{Find github folder}
\usage{
find_github_folder(
  folder_name = jap::default_github_folder(),
  home_dir = jap::default_home_dir()
)
}
\description{
Find github folder
}
