% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_function_list}
\alias{get_function_list}
\title{Get function list from the specified package}
\usage{
get_function_list(package_name, github_name = NA)
}
\arguments{
\item{package_name}{the name of the package}

\item{github_name}{the github account name where the repo is}
}
\value{
function list
}
\description{
Get function list from the specified package
}
\author{
Giovanni Laudanno
}
