% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{install_package}
\alias{install_package}
\title{Install and load a package}
\usage{
install_package(github_name = NA, package_name)
}
\value{
nothing
}
\description{
Install and load a package
}
\author{
Giovanni Laudanno
}
