% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pars_transform_back}
\alias{pars_transform_back}
\title{Transform parameters back}
\usage{
pars_transform_back(pars_transformed)
}
\arguments{
\item{pars_transformed}{parameters of the likelihood functions, transformed
according to y = x / (1 + x)}
}
\value{
the original parameters
}
\description{
Transform parameters back according to x = y / (1 + y)
}
\details{
This is not to be called by the user.
}
