% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{pars_transform_forward}
\alias{pars_transform_forward}
\title{Transform parameters}
\usage{
pars_transform_forward(pars)
}
\arguments{
\item{pars}{vector of parameters:
\itemize{
  \item pars[1] is lambda, the sympatric speciation rate;
  \item pars[2] is mu, the extinction rate;
  \item pars[3] is nu, the multiple allopatric speciation trigger rate;
  \item pars[4] is q, the single-lineage speciation probability;
}}
}
\value{
transformed parameters
}
\description{
Transform parameters according to y = x / (1 + x)
}
\details{
This is not to be called by the user.
}
