% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_functions.R
\name{pn}
\alias{pn}
\title{Pn}
\usage{
pn(lambda, mu, t, n)
}
\arguments{
\item{lambda}{the sympatric speciation rate.}

\item{mu}{the extinction rate.}

\item{t}{time}
}
\value{
pn
}
\description{
Nee's function: pn
}
\author{
Giovanni Laudanno
}
