% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_experiment.R
\name{pocket_experiment}
\alias{pocket_experiment}
\title{Run an experiment on cluster}
\usage{
pocket_experiment(
  account = jap::your_account(),
  projects_folder_name = jap::default_projects_folder(),
  github_name = "Giappo",
  project_name = "sls",
  function_name = "sls_main",
  params,
  cluster_folder = jap::default_cluster_folder(),
  home_dir = jap::default_home_dir(),
  my_email = jap::default_my_email(),
  cluster_partition = "gelifes",
  jap_branch = "master",
  drive = jap::default_drive_choice()
)
}
\arguments{
\item{account}{a peregrine account}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{github_name}{the github account name where the repo is}

\item{project_name}{the name of the project}

\item{function_name}{the name of the function}

\item{params}{parameters for the function in the list format}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{home_dir}{path to the home directory. On windows usually is
a disk like C or D. On mac is always ~.}

\item{cluster_partition}{A string, with name of partition in which to run
the main job. Defaults to \code{"gelifes"}, accepts \code{"regular"},
\code{"short"} and remaining Peregrine partitions.}

\item{drive}{do you want to use google drive to store the data of your
prokects? It can be either TRUE or FALSE}
}
\value{
nothing
}
\description{
Run an experiment on cluster
}
\author{
Giovanni Laudanno
}
