% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocket_experiment.R
\name{pocket_experiment}
\alias{pocket_experiment}
\title{Run an experiment on cluster}
\usage{
pocket_experiment(
  account = jap::your_account(),
  projects_folder_name = "Projects",
  github_name = "Giappo",
  project_name = "sls",
  function_name = "sls_main",
  params,
  drive = FALSE
)
}
\arguments{
\item{account}{a peregrine account}
}
\value{
nothing
}
\description{
Run an experiment on cluster
}
\author{
Giovanni Laudanno
}
