% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_file_exists}
\alias{remote_file_exists}
\title{Check if file exist on cluster}
\usage{
remote_file_exists(file, account = jap::your_account(), session = NA)
}
\arguments{
\item{file}{a file}

\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Check if file exist on cluster
}
\author{
Giovanni Laudanno
}
