% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_install.packages}
\alias{remote_install.packages}
\title{Install packages for a given project}
\usage{
remote_install.packages(
  github_name = NA,
  package_name,
  must_sleep = TRUE,
  account = jap::your_account(),
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Install packages for a given project
}
\author{
Giovanni Laudanno
}
