% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{remote_list.files}
\alias{remote_list.files}
\title{List remote files}
\usage{
remote_list.files(
  dir,
  cluster_folder = jap::default_cluster_folder(),
  account = jap::your_account(),
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
List of files
}
\description{
List remote files
}
\author{
Giovanni Laudanno
}
