% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{run_function}
\alias{run_function}
\title{run a function}
\usage{
run_function(github_name = NA, package_name, function_name,
  fun_arguments)
}
\value{
nothing
}
\description{
run a function
}
\author{
Giovanni Laudanno
}
