% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{run_on_cluster}
\alias{run_on_cluster}
\title{Run a function on cluster}
\usage{
run_on_cluster(
  github_name = NA,
  package_name,
  function_name,
  fun_arguments,
  run_name = "default",
  projects_folder_name = jap::default_projects_folder(),
  cluster_folder = jap::default_cluster_folder(),
  cluster_partition = "gelifes",
  home_dir = jap::default_home_dir(),
  account = jap::your_account(),
  my_email = jap::default_my_email(),
  drive = jap::default_drive_choice(),
  jap_branch = "master",
  session = NA
)
}
\arguments{
\item{github_name}{the github account name where the repo is}

\item{package_name}{the name of the package}

\item{function_name}{the name of the function}

\item{fun_arguments}{the arguments of the function passed either as
list or as string (in the format 'arg1 = x, arg2 = y')}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{cluster_folder}{main folder on the cluster. It can be either 'home'
or 'data'}

\item{cluster_partition}{A string, with name of partition in which to run
the main job. Defaults to \code{"gelifes"}, accepts \code{"regular"},
\code{"short"} and remaining Peregrine partitions.}

\item{home_dir}{path to the home directory. On windows usually is
a disk like C or D. On mac is always ~.}

\item{account}{a peregrine account}

\item{drive}{do you want to use google drive to store the data of your
prokects? It can be either TRUE or FALSE}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Run a function on cluster
}
\author{
Giovanni Laudanno
}
