% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirouette_functions.R
\name{run_pirouette_example}
\alias{run_pirouette_example}
\title{run pirouette example}
\usage{
run_pirouette_example(example_no, account = "p274829", session = NA, gl = TRUE)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
run pirouette example
}
\author{
Giovanni Laudanno
}
