% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{run_project_on_cluster}
\alias{run_project_on_cluster}
\title{run pirouette example}
\usage{
run_project_on_cluster(project_name, function_name, account = "p274829",
  session = NA, fun_arguments)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
run pirouette example
}
\author{
Giovanni Laudanno
}
