% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{run_project_on_cluster}
\alias{run_project_on_cluster}
\title{run pirouette example}
\usage{
run_project_on_cluster(
  project_name,
  function_name,
  account = jap::your_account(),
  session = NA,
  fun_arguments
)
}
\arguments{
\item{project_name}{the name of the project}

\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
NOT WORKING YET
}
\author{
Giovanni Laudanno
}
