% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{upload_cluster_scripts}
\alias{upload_cluster_scripts}
\title{Export cluster scripts}
\usage{
upload_cluster_scripts(
  project_name = "sls",
  account = jap::your_account(),
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Export cluster scripts
}
\author{
Giovanni Laudanno
}
