% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_functions.R
\name{upload_cluster_scripts}
\alias{upload_cluster_scripts}
\title{Export cluster scripts}
\usage{
upload_cluster_scripts(
  project_name = "sls",
  account = jap::your_account(),
  projects_folder_name = jap::default_projects_folder(),
  cluster_folder = jap::default_cluster_folder(),
  home_dir = jap::default_home_dir(),
  drive = jap::default_drive_choice(),
  session = NA
)
}
\arguments{
\item{project_name}{the name of the project}

\item{account}{a peregrine account}

\item{projects_folder_name}{the name you want to give to the folder
containing all your projects}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Export cluster scripts
}
\author{
Giovanni Laudanno
}
