% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh_functions.R
\name{upload_jap_scripts}
\alias{upload_jap_scripts}
\title{Export cluster scripts}
\usage{
upload_jap_scripts(
  account = jap::your_account(),
  cluster_folder = "home",
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Export cluster scripts
}
\author{
Giovanni Laudanno
}
